Subclassed AuthSub variant of GoogleSession
Changed GoogleSession into an abstract class and subclassed the
existing code to handle the auth sub method of authenticating with
google's services. The purpose is to be able to switch out the
production auth sub variant with a client login variant to ease
with integration testing of the service layer.