Refactored ImageRepository
Separated the code of storing and retrieving the images into ImageStore and
ImageRetrieve respectively. This was done because of how divergent the
behaviors are. Storing the images requires uploading the file using the
underlying image service mechanism, E.G. Amazon S3. Retrieving the images
requires simply retrieving the URL of the stored image, and presenting that
to the user of the library.

Furthermore the image retrieval is more or less a database operation, futher
diverging the responsibilities.

I have also added the beginnings of a scaffold to perform integration tests
the live database for the ImageRetrieval object.